Breaking circular dependency by splitting the character mapping off and initializing it in exult.cc.
